[#4803] Reload the Journal details when showing the deliverable name
This caused an odd bug that occurred in Postgres. What would happen is that
the first call would find the Deliverable correctly and set the value/old_value
on the Journal.  But on the second call, the Journal would have the Deliverable
subjects in the value and old_value (as string), which would cause Postgres to
error since it won't allow searching id (int) with the subject (string).

MySQL and SQLite worked because the finder would run but return nil and the next
statements are guarded against nils.